requirements.txt
# Django Related

Django==3.1.6
django-environ==0.4.5
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
django-service-objects==0.7.0
Pillow==8.2.0

# Code formatting

black==20.8b1
isort==5.7.0